Provide option to vary regex
Hello. I mainly wanted to change the name of the string in the middle of the comment so I could have <!-- removeThis --> as well as <!--Deject-->. I think passing a string and building the regex using new RegExp() would be more user friendly but I couldn't get it to work.  Instead, I've just allowed the whole regex to be overwritten.